% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_aws.R
\name{process.campbellsci}
\alias{process.campbellsci}
\title{Process CAMPBELL SCIENTIFIC data.}
\usage{
process.campbellsci(dirAWS, dirCBS)
}
\arguments{
\item{dirAWS}{full path to the directory containing the AWS_DATA folder.\cr}

\item{dirCBS}{full path to the directory CAMPBELL SCIENTIFIC folder on ADT server.}
}
\description{
Read the AWS data from the local folder of CAMPBELL SCIENTIFIC, parse and convert into ADT format.
}
