% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_aws.R
\name{process.pulsonic}
\alias{process.pulsonic}
\title{Process PULSONIC data.}
\usage{
process.pulsonic(dirAWS, dirPLUSO)
}
\arguments{
\item{dirAWS}{full path to the directory containing the AWS_DATA folder.\cr}

\item{dirPLUSO}{full path to the directory PULSONIC folder on ADT server.}
}
\description{
Read the AWS data from the local replicator of PULSONIC, parse and convert into ADT format.
}
