% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_aws.R
\name{process.tahmo}
\alias{process.tahmo}
\title{Process TAHMO data.}
\usage{
process.tahmo(dirAWS, dirTAHMO)
}
\arguments{
\item{dirAWS}{full path to the directory containing the AWS_DATA folder.\cr}

\item{dirTAHMO}{full path to the directory TAHMO folder on ADT server.}
}
\description{
Read the AWS data from the local replicator of TAHMO, parse and convert into ADT format.
}
